pyproject.toml
[project]
name = "chestny-znak-flow-validator"
version = "0.1.0"
description = "Integrity validator for Chestny ZNAK supply chain flows"
authors = [{name = "run-as-daemon.ru"}]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "fastapi>=0.109",
    "pydantic>=1.10",
    "openpyxl>=3.1",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
